requirements.txt
altgraph==0.17
distro==1.4.0
future==0.18.2
numpy==1.18.1
pandas==1.0.1
pefile==2019.4.18
PyInstaller==3.6
PySide2==5.14.1
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
shiboken2==5.14.1
six==1.14.0
tabula-py==2.1.0
